requirements.txt
absl-py==1.4.0
alembic==1.11.1
astunparse==1.6.3
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
deepface==0.0.79
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
gast==0.4.0
gdown==4.7.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.0
python-socketio==5.8.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
simple-websocket==0.10.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
